requirements.txt
tensorflow>=2.12.0
numpy>=1.21.0
Pillow>=9.0.0
scikit-learn>=1.0.0
opencv-python>=4.5.0 